moved graph tests to the proper location
